requirements.txt
Django==2.2.8
django-isbn-field==0.3
djangorestframework==3.11
django-extensions==1.8.1
django-filter==2.2
factory-boy==2.8.1
Faker==0.7.17
psycopg2==2.7.1
